Move NGHTTP2_PRI_* to nghttp2.h, fix too small pri data type in src

+++ src/nghttp2_ssl.cc
@@ -142,7 +142,7 @@ int Spdylay::submit_request(const std::string& scheme,
                             const std::string& hostport,
                             const std::string& path,
                             const std::map<std::string,std::string> &headers,
-                            uint8_t pri,
+                            int32_t pri,
                             const nghttp2_data_provider *data_prd,
                             int64_t data_length,
                             void *stream_user_data)
